requirements.txt
deap==1.4.1
gymnasium==0.29.1
matplotlib==3.6.2
neat_python==0.92
numpy==1.23.4
